Multi-tenency. v.9

+++ src/main/java/org/servantscode/commons/db/ConfigDB.java
@@ -2,6 +2,8 @@ package org.servantscode.commons.db;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.servantscode.commons.search.QueryBuilder;
+import org.servantscode.commons.security.OrganizationContext;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -15,32 +17,28 @@ public class ConfigDB extends DBAccess {
     private static final Logger LOG = LogManager.getLogger(ConfigDB.class);
 
     public String getConfiguration(String config) {
+        QueryBuilder query = select("value").from("configuration").where("config=?", config).inOrg();
         try(Connection conn = getConnection();
-            PreparedStatement stmt = conn.prepareStatement("SELECT value FROM configuration WHERE config=?")) {
+            PreparedStatement stmt = query.prepareStatement(conn);
+            ResultSet rs = stmt.executeQuery()) {
 
-            stmt.setString(1, config);
-            try(ResultSet rs = stmt.executeQuery()) {
-                if(rs.next())
-                    return rs.getString(1);
-            }
+            return rs.next()? rs.getString(1): null;
         } catch (SQLException e) {
             throw new RuntimeException("Could not retrieve configuration property: " + config, e);
         }
-        return null;
     }
 
     public Map<String, String> getConfigurations(String configPrefix) {
+        QueryBuilder query = select("value").from("configuration").where("config LIKE ?", configPrefix + "%").inOrg();
         try(Connection conn = getConnection();
-            PreparedStatement stmt = conn.prepareStatement("SELECT * FROM configuration WHERE config LIKE ?")) {
+            PreparedStatement stmt = query.prepareStatement(conn);
+            ResultSet rs = stmt.executeQuery()) {
 
-            stmt.setString(1, configPrefix + "%");
-            try(ResultSet rs = stmt.executeQuery()) {
-                Map<String, String> results = new HashMap<>();
-                while(rs.next())
-                    results.put(rs.getString("config"), rs.getString("value"));
-                LOG.debug("Retrieved " + results.size() + " properties starting with " + configPrefix);
-                return results;
-            }
+            Map<String, String> results = new HashMap<>();
+            while(rs.next())
+                results.put(rs.getString("config"), rs.getString("value"));
+            LOG.debug("Retrieved " + results.size() + " properties starting with " + configPrefix);
+            return results;
         } catch (SQLException e) {
             throw new RuntimeException("Could not retrieve configuration properties: " + configPrefix, e);
         }
@@ -48,13 +46,14 @@ public class ConfigDB extends DBAccess {
 
     public void patchConfigurations(Map<String, String> configs) {
         try(Connection conn = getConnection();
-            PreparedStatement stmt = conn.prepareStatement("INSERT INTO configuration VALUES(?, ?) " +
-                                                               "ON CONFLICT (config) DO UPDATE SET VALUE=EXCLUDED.value"))
+            PreparedStatement stmt = conn.prepareStatement("INSERT INTO configuration VALUES(?, ?, ?) " +
+                                                               "ON CONFLICT (config, org_id) DO UPDATE SET VALUE=EXCLUDED.value"))
         {
             configs.forEach((config, value) -> {
                     try {
                         stmt.setString(1, config);
                         stmt.setString(2, value);
+                        stmt.setInt(3, OrganizationContext.orgId());
                         stmt.addBatch();
                     } catch (SQLException e) {
                         throw new RuntimeException("Could not patch configuration properties.", e);
@@ -69,11 +68,12 @@ public class ConfigDB extends DBAccess {
 
     public void deleteConfigurations(Set<String> configs) {
         try(Connection conn = getConnection();
-            PreparedStatement stmt = conn.prepareStatement("DELETE FROM configuration WHERE config=?"))
+            PreparedStatement stmt = conn.prepareStatement("DELETE FROM configuration WHERE config=? AND org_id=?"))
         {
             configs.forEach((config) -> {
                 try {
                     stmt.setString(1, config);
+                    stmt.setInt(2, OrganizationContext.orgId());
                     stmt.addBatch();
                 } catch (SQLException e) {
                     throw new RuntimeException("Could not patch configuration properties.", e);
